+++ dash-renderer/src/AppProvider.react.tsx
@@ -1,14 +1,13 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import {Provider} from 'react-redux';
 
 import initializeStore from './store';
 import AppContainer from './AppContainer.react';
 
-import PropTypes from 'prop-types';
-
 const store = initializeStore();
 
-const AppProvider = ({hooks}) => {
+const AppProvider = ({hooks}: any) => {
     return (
         <Provider store={store}>
             <AppContainer hooks={hooks} />
@@ -19,15 +18,17 @@ const AppProvider = ({hooks}) => {
 AppProvider.propTypes = {
     hooks: PropTypes.shape({
         request_pre: PropTypes.func,
-        request_post: PropTypes.func,
-    }),
+        request_post: PropTypes.func
+    })
 };
 
 AppProvider.defaultProps = {
     hooks: {
         request_pre: null,
-        request_post: null,
-    },
+        request_post: null
+    }
 };
 
 export default AppProvider;
+
+